PRODUCTS_REST_API/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
mysql-connector-python==8.0.26
protobuf==4.22.0
pytz==2022.7.1
sqlparse==0.4.3
mysqlclient==2.1.1